исправление замечаний пуллреквеста.
не исправлено:
Работу с консолью (ввод/вывод, System.in/out) можно вынести в отдельный сервис. Так будет удобнее тестировать классы, где этот ввод/вывод используется. Причем System.in/out такому сервису можно будет тоже передать через конструктор. Такой сервис ничего не будет знать о предметной области и может быть переиспользован

Очень длинные строки для чтения, и похоже что для проверки правильного ответа не нужен стрим. Можно попробовать использовать просто какую-нибудь коллекцию. Типа Set. Ну и метод будет гораздо более читаемый, если сначала отобрать правильные ответы в список и дальнейшую работу вести с ним

это поправлю во второй ветке

+++ homework-1/src/main/java/ru/otus/spring/service/AnswerScoring.java
@@ -8,5 +8,5 @@ import java.util.Collection;
  * @author Александр Шабанов
  */
 public interface AnswerScoring {
-  public long getScore(Collection<QuestionAnswerPair> questionAnswerPairCollection);
+  public int getScore(Collection<QuestionAnswerPair> questionAnswerPairCollection);
 }
